menu: Fixed up imports after module rename
Everything is back to working again

Changes to be committed:
	modified:   tps/__main__.py
	modified:   tps/main.py
	modified:   tps/menu.py
	modified:   tps/tph_constants.py
